requirements.txt
tensorflow
pandas 
dvc
notebook
numpy
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .